requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Flask-WTF==1.1.1
WTForms==3.0.1
google-genai
PyMuPDF
python-dotenv
python-docx
fpdf